fix(form): move `FormNumberField` into form namespace
